// src/app/core/api/generated/model/updateEmployeeCommand.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EmployeeContactDto } from './employeeContactDto';
import { EmployeeExperienceDto } from './employeeExperienceDto';
import { EmployeeDocumentDto } from './employeeDocumentDto';
import { EmployeeEducationDto } from './employeeEducationDto';
import { EmployeeAddressDto } from './employeeAddressDto';


export interface UpdateEmployeeCommand { 
    id?: string;
    firstName?: string | null;
    lastName?: string | null;
    arabicName?: string | null;
    birthDate?: string;
    genderId?: number;
    nationalityId?: number;
    maritalStatusId?: number;
    hireDate?: string;
    statusId?: number;
    birthPlace?: string | null;
    religion?: string | null;
    appointmentDecisionNumber?: string | null;
    appointmentDecisionDate?: string | null;
    employmentTypeId?: number | null;
    contacts?: Array<EmployeeContactDto> | null;
    addresses?: Array<EmployeeAddressDto> | null;
    documents?: Array<EmployeeDocumentDto> | null;
    educations?: Array<EmployeeEducationDto> | null;
    experiences?: Array<EmployeeExperienceDto> | null;
}